# contracts/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Bitcoin AI Analyst Backend",
    "version": "2026-09-research-evidence-v2"
  },
  "paths": {
...
    "/api/Alerts": {
      "get": {
        "tags": [
          "Alerts"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "default": "default"
            }
          },
          {
            "name": "unreadOnly",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "take",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          },
          {
            "name": "includeArchived",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Alerts"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "default": "default"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Alerts/unread-count": {
      "get": {
        "tags": [
          "Alerts"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "default": "default"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Alerts/{id}/read": {
      "post": {
        "tags": [
          "Alerts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Alerts/read-all": {
      "post": {
        "tags": [
          "Alerts"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "default": "default"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Alerts/{id}": {
      "delete": {
        "tags": [
          "Alerts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "userId",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "default": "default"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Alerts/deduplicate": {
      "post": {
        "tags": [
          "Alerts"
        ],
        "parameters": [
          {
            "name": "apply",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Confluence/current": {
      "get": {
        "tags": [
          "Confluence"
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "default": "BTCUSDT"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Confluence/history": {
      "get": {
        "tags": [
          "Confluence"
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "default": "BTCUSDT"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Confluence/calculate": {
      "post": {
        "tags": [
          "Confluence"
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "default": "BTCUSDT"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
...
}